+++ kotlin-js/src/jsMain/kotlin/js/promise/PromiseLike.kt
@@ -1,6 +1,17 @@
+@file:Suppress(
+    "WRONG_BODY_OF_EXTERNAL_DECLARATION",
+    "INLINE_EXTERNAL_DECLARATION",
+    "NON_ABSTRACT_MEMBER_OF_EXTERNAL_INTERFACE",
+    "DECLARATION_CANT_BE_INLINED",
+    "NOTHING_TO_INLINE",
+)
+
 package js.promise
 
 import js.errors.JsError
+import js.promise.internal.awaitPromiseLike
+import js.promise.internal.thenToContinuation
+import kotlin.coroutines.Continuation
 
 sealed external interface PromiseLike<out T> :
     PromiseResult<T> {
@@ -23,4 +34,13 @@ sealed external interface PromiseLike<out T> :
         onFulfilled: (T) -> PromiseResult<R>,
         onRejected: (JsError) -> PromiseResult<R>,
     ): PromiseLike<R>
+
+    override suspend inline fun await(): T =
+        awaitPromiseLike(this)
+
+    inline fun thenTo(
+        continuation: Continuation<T>,
+    ) {
+        thenToContinuation(this, continuation)
+    }
 }