+++ kotlin-js/src/jsMain/kotlin/js/temporal/InstantToStringOptions.kt
@@ -1,5 +1,8 @@
 package js.temporal
 
+import kotlinx.js.JsPlainObject
+
+@JsPlainObject
 sealed external interface InstantToStringOptions :
     ToStringPrecisionOptions {
     var timeZone: TimeZoneLike