+++ kotlin-js/src/jsMain/kotlin/js/temporal/ShowCalendarOption.kt
@@ -1,5 +1,8 @@
 package js.temporal
 
+import kotlinx.js.JsPlainObject
+
+@JsPlainObject
 sealed external interface ShowCalendarOption {
     var calendarName: String? /* "auto" | "always" | "never" | "critical" */
 }