+++ kotlin-js/src/jsMain/kotlin/js/temporal/ZonedDateTimeToStringOptions.kt
@@ -1,5 +1,8 @@
 package js.temporal
 
+import kotlinx.js.JsPlainObject
+
+@JsPlainObject
 sealed external interface ZonedDateTimeToStringOptions :
     CalendarTypeToStringOptions {
     var timeZoneName: String? /* "auto" | "never" | "critical" */