+++ kotlin-node/src/jsMain/kotlin/node/fs/exists.kt
@@ -1,7 +1,5 @@
 package node.fs
 
-import js.promise.await
-
 /**
  * Returns `true` if the path exists, `false` otherwise.
  */