+++ kotlin-preact-signals-react/src/jsMain/kotlin/preact/signals/react/useSignalEffect.kt
@@ -1,0 +1,11 @@
+package preact.signals.react
+
+import react.EffectBuilder
+import react.internal.createEffectCallback
+
+fun useSignalEffect(
+    effect: EffectBuilder.() -> Unit,
+) {
+    val callback = createEffectCallback(effect)
+    rawUseSignalEffect(callback)
+}